Fixed design and documentation
